+++ derive_builder/src/options/mod.rs
@@ -65,6 +65,7 @@ pub struct OptionsBuilder<Mode> {
 /// Certain attributes need to be handled differently for `StructOptions` and `FieldOptions`.
 pub trait OptionsBuilderMode: ::std::fmt::Debug {
     fn parse_builder_name(&mut self, lit: &syn::Lit);
+    fn parse_derive(&mut self, nested: &[syn::NestedMetaItem]);
     fn push_deprecation_note<T: Into<String>>(&mut self, x: T) -> &mut Self;
     /// Provide a diagnostic _where_-clause for panics.
     fn where_diagnostics(&self) -> String;
@@ -313,6 +314,9 @@ impl<Mode> OptionsBuilder<Mode>
             },
             "build_fn" => {
                 self.mode.parse_build_fn_options(nested)
+            },
+            "derive" => {
+                self.mode.parse_derive(nested);
             }
             "field" => {
                 self.parse_field_options(nested);