+++ derive_builder/CHANGELOG.md
@@ -8,7 +8,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - customize setter names via `#[builder(setter(name="..."))]`
 - customize build_fn name via `#[builder(build_fn(name="..."))]`
 - suppress build method generation via `#[builder(build_fn(skip))]`
-- set field visibility separate from setter visibility via `#[builder(field(private))]` at the field or struct level
+- derive additional traits via `#[builder(derive(Trait1, Trait2, ...))]`
+- set field visibility separate from setter visibility via
+  `#[builder(field(private))]` at the field or struct level
 
 ### Deprecated
 - Builder fields will no longer be public by default in 0.5.0; relying on this